+++ tests/Models/BulletinTest.php
@@ -15,4 +15,10 @@ class BulletinTest extends PHPUnit_Framework_TestCase {
     {
         $this->assertSame('Hello World', (string)(new Bulletin)->setBody('Hello World'));
     }
+
+    public function testDetailsAreMadeAvailable()
+    {
+        $bulletin = new Bulletin(['Test' => 'Value']);
+        $this->assertSame('Value', $bulletin->getDetail('Test'));
+    }
 }